refactor: move auth to middleware
